refactor out DistributedMergeTree into a different folder
