+++ src/Storages/DistributedMergeTree/tests/gtest_categorize_records_according_to_sequence_range.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 
 #include <DistributedWriteAheadLog/WAL.h>
-#include <Storages/DistributedMergeTreeCallbackData.h>
+#include <Storages/DistributedMergeTree/DistributedMergeTreeCallbackData.h>
 #include <Storages/MergeTree/SequenceInfo.h>
