+++ src/Storages/MergeTree/registerStorageMergeTree.cpp
@@ -4,7 +4,7 @@
 #include <Storages/StorageFactory.h>
 #include <Storages/StorageMergeTree.h>
 #include <Storages/StorageReplicatedMergeTree.h>
-#include <Storages/StorageDistributedMergeTree.h>
+#include <Storages/DistributedMergeTree/StorageDistributedMergeTree.h>
 
 #include <Common/Macros.h>
 #include <Common/OptimizedRegularExpression.h>
